fixed the comment section style, refactored app.js into postpage, still renders like normal
