+++ instagram/src/Components/CommentSection/CommentSection.css
@@ -1,0 +1,36 @@
+.commentSection {
+    display: flex;
+    justify-content: flex-start;
+}
+
+.individualComments {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.commentText {
+    margin-left: 5px;
+}
+
+.username {
+    margin: 10px;
+}
+
+form {
+    display: flex;
+    justify-content: center;
+    align-items: baseline;
+}
+
+form input {
+    width: 70%;
+    margin-bottom: 3px;
+}
+
+form button {
+    width: 20%;
+    border-radius: 6px;
+    margin: 5px;
+    padding: 1px;
+}
