+++ instagram/src/Components/PostContainer/PostContainer.css
@@ -57,7 +57,5 @@ img {
     margin: 0;
 }
 
-.commentSection {
-    display: flex;
-    justify-content: flex-start;
-}
+
+
