+++ instagram/src/Components/SearchBar/SearchBar.css
@@ -18,7 +18,7 @@
 }
 
 .searchBar h1 {
-    margin: 10px 20px;
+    margin: 10px;
     font-size: 1.6rem;
     font-family: 'Dancing Script', cursive;
 }
@@ -30,8 +30,6 @@ input {
 
 .icons {
     display: flex;
-    
-
 }
 
 .icons .fa-compass {
